requirements.txt
Django==5.2.9
djangorestframework==3.16.1
psycopg[binary]==3.3.2
python-dotenv==1.2.1
pytest==9.0.2
pytest-django==4.11.1